ReconstructionByCamera/coltran_obj_pcd/requirements.txt
tensorflow>=2.5.0
numpy
absl-py==0.10.0
tensorflow_datasets
ml_collections
opencv-python
pillow
open3d